Refactor pretalx.common.forms.utils
