+++ src/pretalx/common/forms/validators.py
@@ -1,4 +1,8 @@
+from django import forms
 from django.core.exceptions import ValidationError
+from django.core.validators import MaxValueValidator, MinValueValidator
+from django.utils.formats import date_format
+from django.utils.timezone import get_current_timezone
 from django.utils.translation import gettext_lazy as _
 from zxcvbn import zxcvbn
 
@@ -24,3 +28,49 @@ class ZXCVBNValidator:
         if results.get("score", 0) < self.min_score:
             feedback = ", ".join(results.get("feedback", {}).get("suggestions", []))
             raise ValidationError(_(feedback), params={})
+
+
+class MinDateValidator(MinValueValidator):
+    def __call__(self, value):
+        try:
+            return super().__call__(value)
+        except forms.ValidationError as e:
+            e.params["limit_value"] = date_format(
+                e.params["limit_value"], "SHORT_DATE_FORMAT"
+            )
+            raise e
+
+
+class MinDateTimeValidator(MinValueValidator):
+    def __call__(self, value):
+        try:
+            return super().__call__(value)
+        except forms.ValidationError as e:
+            e.params["limit_value"] = date_format(
+                e.params["limit_value"].astimezone(get_current_timezone()),
+                "SHORT_DATETIME_FORMAT",
+            )
+            raise e
+
+
+class MaxDateValidator(MaxValueValidator):
+    def __call__(self, value):
+        try:
+            return super().__call__(value)
+        except forms.ValidationError as e:
+            e.params["limit_value"] = date_format(
+                e.params["limit_value"], "SHORT_DATE_FORMAT"
+            )
+            raise e
+
+
+class MaxDateTimeValidator(MaxValueValidator):
+    def __call__(self, value):
+        try:
+            return super().__call__(value)
+        except forms.ValidationError as e:
+            e.params["limit_value"] = date_format(
+                e.params["limit_value"].astimezone(get_current_timezone()),
+                "SHORT_DATETIME_FORMAT",
+            )
+            raise e
